% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whois.r
\name{whois}
\alias{whois}
\title{Retrieve WHOIS details for one or more domain names}
\usage{
whois(domains, api_key = whoisxmlapi_key())
}
\arguments{
\item{domains}{character vector of domain names}

\item{api_key}{see \code{\link[=whoisxmlapi_key]{whoisxmlapi_key()}}}
}
\description{
Retrieve WHOIS details for one or more domain names
}
\references{
\url{https://whoisapi.whoisxmlapi.com/docs}
}
