% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoisxmapi-package.r
\docType{package}
\name{whoisxmlapi}
\alias{whoisxmlapi}
\alias{whoisxmlapi-package}
\title{Query 'IP' and Domain Information from 'WhoisXMLAPI'
The 'WhoisXMLAPI' (\url{https://whoisxmlapi.com/}) service provides
current and historical metadata on 'IP' addresses and 'WHOIS' records
for internet domain names. Methods are provided to query and process
query results.}
\description{
Query 'IP' and Domain Information from 'WhoisXMLAPI'
The 'WhoisXMLAPI' (\url{https://whoisxmlapi.com/}) service provides
current and historical metadata on 'IP' addresses and 'WHOIS' records
for internet domain names. Methods are provided to query and process
query results.
}
\author{
Bob Rudis (@hrbrmstr)
}
\keyword{internal}
