% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{whoisxmlapi_key}
\alias{whoisxmlapi_key}
\title{Get or set WHOISXMLAPI_KEY value}
\usage{
whoisxmlapi_key(force = FALSE)
}
\arguments{
\item{force}{Force setting a new WhoisXMLAPI key for the current environment?}
}
\value{
atomic character vector containing the WhoisXMLAPI username
}
\description{
The API wrapper functions in this package all rely on a WhoisXMLAPI
key residing in the environment variable \code{WHOISXMLAPI_KEY}.
The easiest way to accomplish this is to set it
in the \code{.Renviron} file in your home directory.
}
