% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{whoisxmlapi_password}
\alias{whoisxmlapi_password}
\title{Get or set WHOISXMLAPI_PASSWORD value}
\usage{
whoisxmlapi_password(force = FALSE)
}
\arguments{
\item{force}{Force setting a new WhoisXMLAPI password for the current environment?}
}
\value{
atomic character vector containing the WhoisXMLAPI username
}
\description{
The API wrapper functions in this package all rely on a WhoisXMLAPI
username & password residing in the environment variables \code{WHOISXMLAPI_USERNAME}
and \code{WHOISXMLAPI_PASSWORD}. The easiest way to accomplish this is to set it
in the `.Renviron` file in your home directory.
}

