% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account-balance.R
\name{account_balance}
\alias{account_balance}
\title{Retrieve API account balance}
\usage{
account_balance(api_key = whoisxmlapi_key())
}
\arguments{
\item{api_key}{see \code{\link[=whoisxmlapi_key]{whoisxmlapi_key()}}}
}
\description{
Retrieve API account balance
}
\references{
\url{https://whoisapi.whoisxmlapi.com/docs}
}
