% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whois.r
\name{whois}
\alias{whois}
\title{Retrieve WHOIS details for one or more domain names}
\usage{
whois(domains)
}
\arguments{
\item{domains}{character vector of domain names}
}
\value{
nested list of WHOIS information for each domain
}
\description{
Retrieve WHOIS details for one or more domain names
}
\references{
\url{https://www.whoisxmlapi.com/whois-api-doc.php} and
            \url{https://www.whoisxmlapi.com/documentation/whoisapi_documentation/index.html}
}

