% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/whoisxmapi-package.r
\docType{package}
\name{whoisxmlapi}
\alias{whoisxmlapi}
\alias{whoisxmlapi-package}
\title{A package to work with the WhoisXMLAPI <http://whoisxmlapi.com/> service}
\description{
A package to work with the WhoisXMLAPI <http://whoisxmlapi.com/> service
}
\author{
Bob Rudis (@hrbrmstr)
}

